Extract testing of the type system.
